Updates to User, Channel and Thread

+++ restful-web-services/src/main/java/com/sept/rest/webservices/restfulwebservices/resource/ThreadResource.java
@@ -1,68 +1,62 @@
 package com.sept.rest.webservices.restfulwebservices.resource;
 
+import com.sept.rest.webservices.restfulwebservices.repository.ChannelRepository;
 import com.sept.rest.webservices.restfulwebservices.repository.ThreadRepository;
 import com.sept.rest.webservices.restfulwebservices.exception.ThreadNotFoundException;
+import com.sept.rest.webservices.restfulwebservices.model.Channel;
 import com.sept.rest.webservices.restfulwebservices.model.Thread;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 
-import java.util.Arrays;
-import java.util.Date;
 import java.util.List;
+import java.util.Optional;
 
 import javax.management.InvalidAttributeValueException;
-import javax.validation.Valid;
-import org.springframework.beans.factory.annotation.Autowired;
 
 @CrossOrigin(origins="*")
 @RestController
 public class ThreadResource {
 	
 	@Autowired
+	ChannelRepository channelRepository;
+	
+	@Autowired
 	ThreadRepository threadRepository;
 	
 	@GetMapping ("/api/thread")
-	public List<Thread> getAllThreads() {
-		return threadRepository.findAll();
+	public  ResponseEntity<List<Thread>> getAllThreads() {
+		List<Thread> threads = threadRepository.findAll();
+		return new ResponseEntity<>(threads, HttpStatus.OK);
 	}
 	
 	@GetMapping ("/api/thread/{id}")
-	public Thread getThread(@PathVariable(name = "id") Long id) throws ThreadNotFoundException {
-		return threadRepository.findById(id)
-				.orElseThrow(() -> new ThreadNotFoundException(id));
+	public ResponseEntity<Thread>  getThread(@PathVariable Long id) throws ThreadNotFoundException {
+		Thread threads = threadRepository.findById(id).orElseThrow(() -> new ThreadNotFoundException(id));
+		return new ResponseEntity<>(threads, HttpStatus.OK);
 	}
 	
-	@PostMapping("/api/thread")
-    public Thread createThread(@RequestBody Thread thread) {
-		thread.setDatetime(new Date().toString());
-        return threadRepository.save(thread);
-    }
-	
 	@DeleteMapping("/api/thread/{id}")
-	public ResponseEntity<?> deleteThread(@PathVariable(name = "id") Long id) throws ThreadNotFoundException {
+	public ResponseEntity<?> deleteThread(@PathVariable Long id) throws ThreadNotFoundException {
 		Thread thread = threadRepository.findById(id)
 				.orElseThrow(() -> new ThreadNotFoundException(id));
-		threadRepository.delete(thread);
-		return ResponseEntity.ok().build();
+		Optional<Channel> primaryChannel = channelRepository.findById(thread.getPrimaryChannel());
+		
+		if (primaryChannel.isPresent()) {
+			primaryChannel.get().removeThread(thread);
+			threadRepository.delete(thread);
+		}
+		
+		return new ResponseEntity<>(HttpStatus.OK);
 	}
 	
 	@PutMapping("/api/thread/{id}")
-    public Thread updateNote(@PathVariable(value = "id") Long id,
-                           @Valid @RequestBody Thread threadDetails) throws ThreadNotFoundException, InvalidAttributeValueException {
+    public ResponseEntity<?> updateNote(@PathVariable Long id,
+    		@RequestBody Thread thread) throws ThreadNotFoundException, InvalidAttributeValueException {
+		thread.setId(id);
+        threadRepository.save(thread);
 
-		Thread thread = threadRepository.findById(id)
-                .orElseThrow(() -> new ThreadNotFoundException(id));
-
-		thread.setTitle(threadDetails.getTitle());
-        thread.setContent(threadDetails.getContent());
-        thread.setDatetime(threadDetails.getDatetime());
-        thread.setArchived(threadDetails.isArchived());
-        thread.setUpspikes(threadDetails.getUpspikes());
-        thread.setDownspikes(threadDetails.getDownspikes());
-
-        Thread updatedBook = threadRepository.save(thread);
-
-        return updatedBook;
+		return new ResponseEntity<>(HttpStatus.OK);
     }
 }
